Add debug flag to all commands
I wanted to add the `debug` flag to the `git.Run` function, so each
command that calls that function (basically all of them) needed to parse
the `debug` flag. This brings up the need for a utility function that
parses for the `debug` flag since it's a lot of repeated code.

This change also fixes the issue where if the only arg is a gitArg, then
the nothing gets passed to the underlying `git` command.